pyproject.toml
[project]
name = "memory-agent"
version = "0.1.0"
description = "Conversational agent with persistent memory using hybrid search (vector + BM25)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "google-genai>=1.0.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.0.0",
]
